+++ compliance_tests/TestIso_8_5_7.cpp
@@ -90,7 +90,7 @@ class TestIso_8_5_7 : public test_lib::TestBase
         {
             SetupTestEnvironment();
 
-            for (int test_variant = 0; test_variant < test_variants.size(); test_variant++)
+            for (size_t test_variant = 0; test_variant < test_variants.size(); test_variant++)
             {
                 PrintVariantInfo(test_variants[test_variant]);
